refactor: reorganize runtime layout with langserve and mcp
