+++ docs/operations/docker/OPTIMIZATION_SUMMARY.md
