+++ docs/reports/AUTOMATION_OPTIMIZATION_REPORT.md
