+++ docs/reports/INTELLIGENT_CACHING_REPORT.md
