+++ security-agent-system/apps/__init__.py
@@ -1,0 +1,1 @@
+"""Application entrypoints for the Security Agent System."""
