+++ security-agent-system/apps/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Command line interface for orchestrating the platform."""
+
+from .main import cli
+
+__all__ = ["cli"]
