+++ security-agent-system/apps/langserve/__init__.py
@@ -1,0 +1,1 @@
+"""LangServe deployment for the Security Agent System."""
