+++ security-agent-system/apps/mcp/__init__.py
@@ -1,0 +1,1 @@
+"""Model Context Protocol server integration for the platform."""
