+++ security-agent-system/requirements.txt
@@ -29,6 +29,7 @@ chromadb==0.4.22
 fastapi==0.109.2
 uvicorn==0.27.1
 httpx==0.26.0
+langserve==0.1.0
 
 # Async Support
 aiohttp==3.9.3
@@ -40,6 +41,7 @@ numpy==1.26.3
 # Integration
 slack-sdk==3.26.2
 prometheus-client==0.19.0
+mcp==0.1.0
 
 # Security
 cryptography==42.0.2
