+++ security-agent-system/security_agent_system/__init__.py
@@ -1,0 +1,5 @@
+"""Security Agent System package exposing agents, workflows, and runtimes."""
+
+from .core.config import settings  # noqa: F401
+
+__all__ = ["settings"]
