+++ security-agent-system/security_agent_system/infrastructure/vector_db.py
